Move Backend API Provider & Context to shared dir
