+++ frontend/benefit/applicant/src/pages/_app.tsx
@@ -4,13 +4,12 @@ import { appWithTranslation } from 'benefit/applicant/i18n';
 import { AppProps } from 'next/app';
 import React from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
+import BackendAPIProvider from 'shared/backend-api/BackendAPIProvider';
 import Content from 'shared/components/content/Content';
 import Layout from 'shared/components/layout/Layout';
 import GlobalStyling from 'shared/styles/globalStyling';
 import theme from 'shared/styles/theme';
 import { ThemeProvider } from 'styled-components';
-
-import BackendAPIProvider from '../backend-api/BackendAPIProvider';
 
 const queryClient = new QueryClient();
