+++ frontend/kesaseteli/employer/src/__tests__/utils/components/render-component.tsx
@@ -1,9 +1,9 @@
 import AxiosTestContext from 'kesaseteli/employer/__tests__/utils/backend/axios-test-context';
-import BackendAPIContext from 'kesaseteli/employer/backend-api/BackendAPIContext';
 import { NextRouter } from 'next/router';
 import React from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import getDefaultReactQueryTestClient from 'shared/__tests__/utils/react-query/get-default-react-query-test-client';
+import BackendAPIContext from 'shared/backend-api/BackendAPIContext';
 import { render, RenderResult } from 'test-utils';
 
 const renderComponent = (
