--- frontend/kesaseteli/employer/src/backend-api/BackendAPIContext.tsx
@@ -1,4 +1,0 @@
-import { AxiosInstance } from 'axios';
-import React from 'react';
-
-export default React.createContext<AxiosInstance | null>(null);
